+++ offline-1/src/conc_scene/rolling_ball_scene.cpp
@@ -35,9 +35,9 @@ rolling_ball_scene::rolling_ball_scene()
     transform left_wall_transform(vector3(5.0f, 0.25f, 0.0f));
     transform right_wall_transform(vector3(-5.0f, 0.25f, 0.0f));
     transform top_wall_transform(vector3(0.0f, 0.25f, 4.9f),
-        vector3(0.0f, (float)M_PI / 2.0f, 0.0f));
+        vector3(0.0f, M_PI / 2.0f, 0.0f));
     transform bot_wall_transform(vector3(0.0f, 0.25f, -4.9f),
-        vector3(0.0f, (float)M_PI / 2.0f, 0.0f));
+        vector3(0.0f, M_PI / 2.0f, 0.0f));
     m_left_wall_collider = new box_collider(vector3(0.2f, 0.5f, 10.0f));
     m_right_wall_collider = new box_collider(vector3(0.2f, 0.5f, 10.0f));
     m_top_wall_collider = new box_collider(vector3(0.2f, 0.5f, 9.8f));
